Converted osu commands to cogs
